Move pagination test to new test file
